[ refacetor ] query 관련 커스텀훅 별개로 분리
